requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-pubsub
pydantic
google-cloud-bigquery
sendgrid
